# EMCSQL/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api Facturacion EMC",
    "version": "v1"
  },
  "paths": {
...
    "/api/FacturaVenta/PostFacturaVenta": {
      "post": {
        "tags": [
          "FacturaVenta"
        ],
        "operationId": "PostFacturaVenta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaVentaDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaVentaDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaVentaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/FacturaVenta/GetFacturasVentasFechas/{idEmpresa}/{fechaInicial}/{fechaFinal}": {
      "get": {
        "tags": [
          "FacturaVenta"
        ],
        "operationId": "GetFacturasVentasFechas",
        "parameters": [
          {
            "name": "idEmpresa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fechaInicial",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "fechaFinal",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaVentaDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaVentaDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaVentaDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/FacturaVenta/GetFacturasVentas/{idEmpresa}": {
      "get": {
        "tags": [
          "FacturaVenta"
        ],
        "operationId": "GetFacturasVentas",
        "parameters": [
          {
            "name": "idEmpresa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaVentaDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaVentaDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaVentaDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/FacturaVenta/GetNumeroFactura/{idEmpresa}": {
      "get": {
        "tags": [
          "FacturaVenta"
        ],
        "operationId": "GetNumeroFactura",
        "parameters": [
          {
            "name": "idEmpresa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InicioFacturacionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InicioFacturacionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InicioFacturacionDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "mensaje": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FacturaVentaDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "numeroFactura": {
            "type": "string",
            "nullable": true
          },
          "idEmpresa": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "rfc": {
            "type": "string",
            "nullable": true
          },
          "idCliente": {
            "type": "integer",
            "format": "int32"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "porcentajeImpuesto": {
            "type": "number",
            "format": "double"
          },
          "impuesto": {
            "type": "number",
            "format": "double"
          },
          "porcentajeDescuento": {
            "type": "number",
            "format": "double"
          },
          "descuento": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "fechaModificacion": {
            "type": "string",
            "format": "date-time"
          },
          "idUsuario": {
            "type": "integer",
            "format": "int32"
          },
          "activo": {
            "type": "boolean"
          },
          "idMetodoPago": {
            "type": "integer",
            "format": "int32"
          },
          "metodoPago": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InicioFacturacionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idEmpresa": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "activo": {
            "type": "boolean"
          },
          "numeroFactura": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}